+++ robot_state.py
@@ -7,12 +7,9 @@ from wpilib import DataLogManager, DriverStation, Field2d, SmartDashboard, Mecha
 from wpimath.geometry import Pose2d
 from wpimath.kinematics import ChassisSpeeds, SwerveModuleState
 
-from constants import Constants
-from lib.limelight import LimelightHelpers, PoseEstimate
 from subsystems.elevator import ElevatorSubsystem
 from subsystems.pivot import PivotSubsystem
 from subsystems.swerve import SwerveSubsystem
-from subsystems.vision import Limelight
 
 
 class RobotState:
@@ -93,57 +90,5 @@ class RobotState:
         """Returns the current pose of the robot on the field (blue-side origin)."""
         return self._swerve.get_state().pose
 
-    def should_use_vision_measurements(self) -> bool:
-        return abs(self._swerve.pigeon2.get_angular_velocity_z_world().value) <= 720
-    
-    def add_vision_measurements(self, limelight: Limelight, estimate: PoseEstimate, standard_deviation: tuple[float, float, float] = (0.7, 0.7, 9999999)) -> None:
-        """
-        Adds MegaTag2 Pose Estimates from the given limelight.
-
-        Modified from Limelight's MegaTag 2 documentation, which can be found at
-        https://docs.limelightvision.io/docs/docs-limelight/pipeline-apriltag/apriltag-robot-localization-megatag2
-
-        :param standard_deviation: The standard deviations of the pose estimate. Lower values tell the pose estimator to trust the pose estimate more
-        :type standard_deviation: tuple[float, float, float]
-        :rtype: None
-        """
-        if limelight == Constants.VisionConstants.FRONT_CENTER:
-            LimelightHelpers.set_imu_mode(limelight.name(), 1)
-
-        LimelightHelpers.set_robot_orientation(
-            limelight.name(),
-            self._swerve.pigeon2.get_yaw().value,
-            self._swerve.pigeon2.get_angular_velocity_z_world().value,
-            self._swerve.pigeon2.get_pitch().value,
-            self._swerve.pigeon2.get_angular_velocity_y_world().value,
-            self._swerve.pigeon2.get_roll().value,
-            self._swerve.pigeon2.get_angular_velocity_x_world().value
-        )
-
-        if abs(self._swerve.pigeon2.get_angular_velocity_z_world().value) <= 720 and estimate.tag_count > 0:
-            self._swerve.set_vision_measurement_std_devs(standard_deviation)
-            self._swerve.add_vision_measurement(estimate.pose, utils.fpga_to_current_time(estimate.timestamp_seconds))
-
     def should_pivot_move(self) -> bool:
         return self._elevator.is_at_setpoint()
-
-    def get_latency_compensated_pose(self, dt: float) -> Pose2d:
-        """Returns the current pose of the robot on the field (blue-side origin),
-        compensated for latency.
-
-        :param dt: The amount of time in seconds since the last
-            update.
-        :type dt: float
-        :return: The current pose of the robot on the field with
-            latency compensation.
-        :rtype: Pose2d
-        """
-        state = self._swerve.get_state()
-        speeds = state.speeds
-        pose = state.pose
-
-        return Pose2d(
-            pose.X() + speeds.vx * dt,
-            pose.Y() + speeds.vy * dt,
-            pose.rotation() + speeds.omega * dt
-            )
